Reorg some files for clarity
