+++ frontend/src/components/Sidebar/ActiveWorkspaces/index.jsx
@@ -5,7 +5,7 @@ import "react-loading-skeleton/dist/skeleton.css";
 import Workspace from "../../../models/workspace";
 import ManageWorkspace, {
   useManageWorkspaceModal,
-} from "../../Modals/ManageWorkspace";
+} from "../../Modals/MangeWorkspace";
 import paths from "../../../utils/paths";
 import { useParams } from "react-router-dom";
 
@@ -51,11 +51,10 @@ export default function ActiveWorkspaces() {
           >
             <a
               href={isActive ? null : paths.workspace.chat(workspace.slug)}
-              className={`flex flex-grow w-[75%] h-[36px] gap-x-2 py-[5px] px-4 border border-slate-400 rounded-lg text-slate-800 dark:text-slate-200 justify-start items-center ${
-                isActive
+              className={`flex flex-grow w-[75%] h-[36px] gap-x-2 py-[5px] px-4 border border-slate-400 rounded-lg text-slate-800 dark:text-slate-200 justify-start items-center ${isActive
                   ? "bg-gray-100 dark:bg-stone-600"
                   : "hover:bg-slate-100 dark:hover:bg-stone-900 "
-              }`}
+                }`}
             >
               <Book className="h-4 w-4" />
               <p className="text-slate-800 dark:text-slate-200 text-xs leading-loose font-semibold">
